refactor: add db init in Model.go
